clean up JSON generation for i18n strings
