+++ scripts/generate-locales.ts
@@ -1,3 +1,3 @@
-import { generateLocales } from '../utilities/generate-locales';
+import { generateLocales } from '../src/lib/i18n/generate-locales';
 
 await generateLocales();
